PEAR-2695 - linting, doc cleanup, etc
